feat: mail authentication & redis
mail 인증에 대한 구현 완료
local 환경에서 redis 연동 완료

exception JSON 처리 및 code 추가 필요
Cloud 위에서의 redis 연동 필요

+++ backend/src/main/java/com/smusoak/restapi/redis/RedisService.java
@@ -43,7 +43,7 @@ public class RedisService {
         redisTemplate.delete(key);
     }
 
-    public void setExpire(String key, int ttl) {
+    public void setExpire(String key, long ttl) {
         redisTemplate.expire(key, ttl, TimeUnit.MILLISECONDS);
     }
